# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-tenant truck management system with real-time tracking",
        "title": "Truck Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/cargo": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all cargo shipments for the company with filtering options",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargo"
                ],
                "summary": "Get cargo list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by cargo type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by priority",
                        "name": "priority",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by truck ID",
                        "name": "truck_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter assigned/unassigned cargo",
                        "name": "assigned",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search in title, tracking number, or description",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new cargo shipment to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargo"
                ],
                "summary": "Create a new cargo",
                "parameters": [
                    {
                        "description": "Cargo data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateCargoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Cargo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/cargo/track/{tracking_number}": {
            "get": {
                "description": "Retrieve cargo information using tracking number",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargo"
                ],
                "summary": "Get cargo by tracking number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tracking Number",
                        "name": "tracking_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Cargo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/cargo/track/{tracking_number}/details": {
            "get": {
                "description": "Get comprehensive tracking information including progress and ETA",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargo"
                ],
                "summary": "Get detailed cargo tracking info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tracking Number",
                        "name": "tracking_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CargoTrackingResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/cargo/trucks/{truck_id}/cargo": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all cargo shipments assigned to a specific truck",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargo"
                ],
                "summary": "Get cargo assigned to truck",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Truck ID",
                        "name": "truck_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Cargo"
                            }
                        }
                    }
                }
            }
        },
        "/cargo/trucks/{truck_id}/nearby-cargo": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get cargo near truck's current location for efficient pickup routing",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargo"
                ],
                "summary": "Get nearby cargo for pickup",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Truck ID",
                        "name": "truck_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "default": 10,
                        "description": "Search radius in kilometers",
                        "name": "radius",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Cargo"
                            }
                        }
                    }
                }
            }
        },
        "/cargo/unassigned": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all cargo that hasn't been assigned to any truck",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargo"
                ],
                "summary": "Get unassigned cargo",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Cargo"
                            }
                        }
                    }
                }
            }
        },
        "/cargo/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a specific cargo shipment by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargo"
                ],
                "summary": "Get a cargo by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cargo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Cargo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update cargo shipment information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargo"
                ],
                "summary": "Update a cargo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cargo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated cargo data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateCargoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Cargo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a cargo shipment from the system",
                "tags": [
                    "cargo"
                ],
                "summary": "Delete a cargo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cargo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/cargo/{id}/assign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Assign a cargo shipment to a specific truck",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargo"
                ],
                "summary": "Assign cargo to truck",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cargo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Assignment data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AssignCargoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Cargo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/cargo/{id}/route": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the complete route history with GPS coordinates",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cargo"
                ],
                "summary": "Get cargo route history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cargo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CargoEvent"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AssignCargoRequest": {
            "type": "object",
            "required": [
                "truck_id"
            ],
            "properties": {
                "truck_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.CargoTrackingResponse": {
            "type": "object",
            "properties": {
                "cargo": {
                    "$ref": "#/definitions/models.Cargo"
                },
                "current_status": {
                    "type": "string"
                },
                "estimated_eta": {
                    "type": "string"
                },
                "last_update": {
                    "type": "string"
                },
                "progress": {
                    "description": "0-100%",
                    "type": "number"
                },
                "recent_events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CargoEvent"
                    }
                },
                "truck_location": {
                    "$ref": "#/definitions/models.TruckLocation"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}